refactor add module button
